+++ src/test/java/racingcar/domain/car/CarNameTest.java
@@ -1,4 +1,4 @@
-package racingcar.domain.game.car;
+package racingcar.domain.car;
 
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
@@ -6,7 +6,7 @@ import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.ValueSource;
-import racingcar.exception.Err;
+import racingcar.util.Err;
 
 class CarNameTest {
