+++ src/test/java/racingcar/domain/car/CarTest.java
@@ -1,4 +1,4 @@
-package racingcar.domain.game.car;
+package racingcar.domain.car;
 
 import static org.assertj.core.api.Assertions.assertThat;
