+++ src/test/java/racingcar/domain/car/FixedMoveStrategy.java
@@ -1,4 +1,4 @@
-package racingcar.domain.game.car;
+package racingcar.domain.car;
 
 import racingcar.move.MoveStrategy;
