+++ src/test/java/racingcar/validator/InputValidatorTest.java
@@ -6,7 +6,7 @@ import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.ValueSource;
-import racingcar.exception.Err;
+import racingcar.util.Err;
 
 class InputValidatorTest {
     private InputValidator inputValidator;
